virtual_network: Add hotplug test for vdpa interface
This PR enhances hotplug test for vdpa interface and
moves scripts to virtual_network dir.
