--- libvirt/tests/cfg/virtual_interface/interface_hotplug.cfg
@@ -1,29 +1,0 @@
-- interface_hotplug:
-    type = interface_hotplug
-    start_vm = no
-    repeat_times = 1
-
-    variants dev_type:
-        - vdpa:
-            only x86_64
-            func_supported_since_libvirt_ver = (7, 3, 0)
-            func_supported_since_qemu_kvm_ver = (6, 0, 0)
-            vdpa_dev = "vdpa0"
-            iface_dict = {"source": {'dev':'/dev/vhost-vdpa-0'}}
-            variants test_target:
-                - simulator:
-                - mellanox:
-            variants:
-                - @default:
-                - driver_queues:
-                    only mellanox
-                    driver_queues = "8"
-                    vdpa_mgmt_tool_extra = "max_vqp ${driver_queues}"
-                    iface_dict = {'model': 'virtio', 'source': {'dev': '/dev/vhost-vdpa-0'}, 'type_name': 'vdpa', 'driver': {'driver_attr': {'queues': ${driver_queues}}}, 'mac_address': mac_addr}
-                    required_kernel = [5.14.0,)
-                    func_supported_since_libvirt_ver = (8, 0, 0)
-                    variants:
-                        - vcpu_gt_queues:
-                            vm_attrs = {'vcpu': 4}
-                        - vcpu_lt_queues:
-                            vm_attrs = {'vcpu': 16}
