+++ libvirt/tests/cfg/virtual_network/hotplug/attach_detach_device/hotplug_hotunplug_vdpa_interface.cfg
@@ -1,0 +1,29 @@
+- virtual_network.hotplug.vdpa_interface:
+    type = hotplug_hotunplug_vdpa_interface
+    start_vm = no
+    func_supported_since_libvirt_ver = (7, 3, 0)
+    func_supported_since_qemu_kvm_ver = (6, 0, 0)
+    vdpa_dev = "vdpa0"
+    iface_dict = {"source": {'dev':'/dev/vhost-vdpa-0'}}
+    only x86_64
+
+    variants test_target:
+        - simulator:
+        - mellanox:
+    variants:
+        - @default:
+        - driver_queues:
+            only mellanox
+            driver_queues = "8"
+            vdpa_mgmt_tool_extra = "max_vqp ${driver_queues}"
+            iface_dict = {'model': 'virtio', 'source': {'dev': '/dev/vhost-vdpa-0'}, 'type_name': 'vdpa', 'driver': {'driver_attr': {'queues': ${driver_queues}}}, 'mac_address': mac_addr}
+            required_kernel = [5.14.0,)
+            func_supported_since_libvirt_ver = (8, 0, 0)
+            variants:
+                - vcpu_gt_queues:
+                    vm_attrs = {'vcpu': 4}
+                - vcpu_lt_queues:
+                    vm_attrs = {'vcpu': 16}
+    variants test_scenario:
+        - hotplug_hotunplug:
+        - hotunplug:
